requirements_advanced.txt
# Core RAG Dependencies
langchain==0.1.0
langchain-community==0.0.13
langchain-openai==0.0.6
langchain-pinecone==0.0.3
python-dotenv==1.0.0
openai==1.12.0
pinecone-client==3.0.0

# Evaluation Framework
pandas==2.2.0
numpy==1.26.3

# Streamlit Dashboard
streamlit==1.31.0
plotly==5.18.0

# Testing and Quality
pytest==8.0.0
black==24.1.0
flake8==7.0.0